move parallaxsprite to its own file and give it type definitions
