+++ src/@types/lib/parallax-sprite.d.ts
@@ -1,0 +1,12 @@
+declare module ParallaxSprite {
+  export type LayersConfig = LayerConfig[];
+
+  export type LayerConfig = {
+    key: string;
+  };
+
+  export type Layer = {
+    tileSprite: Phaser.GameObjects.TileSprite;
+    scrollFactor: number;
+  }
+}
